Initial set of architectural decisions
Why these changes are being introduced:

We want to record the decisions we make about this system.

Relevant ticket(s):

N/A.

How this addresses that need:

This adds a few initial ADRs:

* The decision to use ADRs.
* The decision to use Rails.
* The decision to implement a GraphQL API.
* The decision to use a monolithic architecture.

Side effects of this change:

It may be of future interest to know that ADR 0003 began with
the assumption that we would implement a REST API. After some
team discussion, we decided to use GraphQL instead.

co-authored-by: Jeremy Prevost <JPrevost@users.noreply.github.com>
co-authored-by: Matt Bernhardt <mjbernha@mit.edu>