Implement initial GraphQL interface
Why these changes are being introduced:

We need a GraphQL interface to record SearchEvents from TACOS
consumers.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-234

How this addresses that need:

This adds a SearchEvent type and corresponding query field that
returns information about the search event and creates a new
SearchEvent (and Term, if applicable).

Side effects of this change:

* Removed a couple of generated GraphQL test fields.
* Miscellaneous rubocop fixes.