Adds standard identifiers detection
Why are these changes being introduced:

Detection of patterns will eventually help us with categorizing the
incoming search terms, which is the ultimate goal of this application.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ENGX-242
* https://mitlibraries.atlassian.net/browse/ENGX-241

The StandardIdentifier patterns and tests were originally developed as
part of our TIMDEX UI application and have been extracted and modified
for use in this application.

https://github.com/MITLibraries/timdex-ui

co-authored-by: Matt Bernhardt <mjbernha@mit.edu>