Add Heroku config
Why these changes are being introduced:

We need a Procfile and app.json for our Heroku config.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-239

How this addresses that need:

This adds a Procfile and app.json with the initial settings we'll
need.

Side effects of this change:

The pg has been added to production.