Add ADR to select an authorization gem
Why these changes are being introduced:

We will need authorization in TACOS for the staff UI, which
first requires us to choose an authorization gem.

Relevant ticket(s):

* [TCO-30](https://mitlibraries.atlassian.net/browse/TCO-30)

How this addresses that need:

This adds an ADR to select an authorization gem from among
CanCanCan, Pundit, and Action Policy.

Side effects of this change:

The decision we make will affect how we implement authorization.