Adds CSV loader for SearchEvents
https://mitlibraries.atlassian.net/browse/TCO-35

This loader is not meant to be used in production, but instead be a
convenience for loading data into development environments to allow for
easier development of detection algorithms or reporting.

There is a Dataclip that exports data in the expected format, but it
is also easy to generate your own file as it is just `term` and
`timestamp` that is expected.

Currently, `source` is provided via CLI but if we find it would be
better to include the source in the CSV that would be an easy change
to make.