Add ADR to select an authentication approach
Why these changes are being introduced:

We will need authentication in TACOS for the staff UI.

Relevant ticket(s):

* [TCO-28](https://mitlibraries.atlassian.net/browse/TCO-28)

How this addresses that need:

This adds an ADR to select an authentication approach from among
Devise, Clearance, Sorcery, AuthLogic, and a custom solution using
built-in Rails security features.

Side effects of this change:

We ended up with two ADRs with the same number. Those have been
renumbered for clarity.