Add Devise and Omniauth for authentication
Why these changes are being introduced:

We will need user authentication on this app, as some features will
be for authorized staff use only.

Relevant ticket(s):

* [TCO-29](https://mitlibraries.atlassian.net/browse/TCO-29)

How this addresses that need:

This adds and configures Devise, Omniauth, and the Omniauth OIDC
strategy, so we can authenticate with Touchstone. Devise requires
a minimal User model, to which I've added an email field and
some validations.

This commit does not include fake auth for local dev and PR builds.
That feature will come in a subsequent commit.

Side effects of this change:

* The `sign_out` route is defined manually, as we are not using the
`database_authenticatable` module that provides additional routes
(including `destroy_user_session_path`).
* This initial setup is pretty spare. We will likely need to add
some more fields to the User model, and perhaps further Omniauth
config as we continue to learn about Okta.
* The sign-in/sign-out links are now buttons because Rails has
dropped support for `method: :post`. It seems that
[Turbo can replicate this behavior](https://discuss.rubyonrails.org/t/rails-7-0-4-link-to-method-post-no-longer-works/82321),
but I was still seeing GET requests with Turbo. I'd still prefer to
use links, but figuring out how to get them to work felt beyond the
scope of this ticket.