Adds exact journal matches to historical analysis
Why are these changes being introduced:

* Understanding our ability to detect search intent over time is
  core to our ability to know how we are doing

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TCO-54

How does this address that need:

* Adds new field to track exact journal matches to Metrics::Algorithms
* Updates Metrics::Algorithms to run journal exact matches in addition
  to the existing StandardIdentifer matches. This included a refactor
  to better support multiple match types.

Document any side effects to this change:

Journal exact matching is not guaranteed to be an indicator of user
search intent because Journal names are also common words in many cases.
When we build our our validation workflows, we'll be able to understand
what percentage of these types of matches are definitely Journals and
what percentage is ambiguous. We can likey update our algorithm to drop
some of the more ambiguous detections at that point.