Add fake authentication for development
Why these changes are being introduced:

We need an easy way to bypass authentication in local dev
and PR builds.

Relevant ticket(s):

* [TCO-29](https://mitlibraries.atlassian.net/browse/TCO-29)

How this addresses that need:

This implements fake auth in a similar way as we've done in ETD,
using an initializer that checks the fake auth status as evaluated
by a FakeAuthConfig module.

Side effects of this change:

We're missing an opportunity to use Keycloak for local authentication,
which would better simulate the Touchstone process. That may be
worth revisiting at some point.