Rubocop: add todolist
We have a fair amount of not-autofixable code we have already accepted into the repository.
Rather than failing on the same code each time we run the linters, we can note which code is
known to be not following our standards and only fail the linters on changed code.

`bundle exec rubocop --auto-gen-config` is the way to do that.