Expand Categorization to include SuggestedResource
This is not as efficient as it could be. Because the Detector model
only stores a boolean, rather than a pointer to which SR record was
detected, we have to re-run the full_term_match method and look
up the relevant category that way.

This is a prototype, so hopefully this is okay for now.