Refactor detector GraphQL and implement journal detection type
Why these changes are being introduced:

We've implemented a journal detection model but not the corresponding GraphQL. As this will be the second of many
detectors, it also makes sense to reconsider how the detector GraphQL is modeled.

Relevant ticket(s):

* [TCO-50](https://mitlibraries.atlassian.net/browse/TCO-50)

How this addresses that need:

This makes `journals` and `standard_identifiers` fields within a new `DetectorsType`. Much like the `Detector` module,
this new type is not to any ActiveRecord models and acts more as a namespace to contain the various detectors.

Side effects of this change:

* The graphql-ruby docs [advise against overusing JSON](https://graphql-ruby.org/api-doc/2.3.14/GraphQL/Types/JSON.html)
as a GraphQL type, but this feels like a good use case for it.
* I'm not certain whether my solution to the `detectors` field is idiomatic. It feels particularly odd given that,
in the SearchEventType, there is a method for the `phrase` field that returns the same thing.
* The gitignore file has been updated to ignore Lando config. This ended up being unrelated to this changeset, but it
might be useful during cassette regeneration.