Document final decision with ADR, and cleanup
** Why are these changes being introduced:

After the final set of options was discussed, the team decided which
option we would pursue for the time being. This decision needed to be
documented in a clear way.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-69

** How does this address that need:

This combines all the individual class diagrams into an ADR document,
which summarizes each option - including the principle diagram which
includes its key characteristics. This document summarizes our decisioin
to pursue prototype D.

The original class diagram is updated to reflect only this selected
approach.

Additionally, the workflow is summarized in a new explanation document
of the Categorization Workflow.

** Document any side effects to this change:

None in this commit (the first commit in this PR still includes a small
wording change to the Work Activity Analysis).