Implement Detector, Category, DetectorCategory
** Why are these changes being introduced:

Our chosen architecture calls for a set of models that will comprise a
sort of "knowledge graph", which TACOS will consult during the
categorization process. This includes classes for Category, Detector,
and a linking DetectorCategory class. The Detector and DetectorCategory
classes will each define a confidence value.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-82

** How does this address that need:

This defines those classes. The migration includes the creation of the
needed records for each class: Three category records, six detectors,
and five records which link between them.

There are currently no detectors which map to two of the categories,
although we have talked about those being needed. Additionally, one
detector, SuggestedResource, is unique in that specific records will
count toward each category - so it isn't appropriate to have a link
record which uniformly connects to only one category.

** Document any side effects to this change:

We previously had a Detector model file, but it was set as a Module in
order to provide a namespace for its subclasses. This has been updated
to be just a class, which impacts the dashboard and test files. Also,
there was a method to define a table name prefix, which has been moved
from the Detector file to the subclass files.