Customize administrate dashboards
** Why are these changes being introduced:

The dashbaords for detectors, categories, and detector_categories need
to be customized to make them useful.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-85

** How does this address that need:

This does a few different things:

* Ensures that the dashboards for Detectors, Categories, and
  DetectorCategories appear in the administrate sidebar. These links are
  in a cluster, and I've added code comments to the routes file to
  clarify the routes (these clusters may also inform our later roles
  work)

* Customizes the dashboards for these resources, positioning the name
  fields more prominently. The display text for these resources is also
  customized, using names where possible (including, for the linking
  class, the name of the two records being linked).

** Document any side effects to this change:

Hopefully having the second commit in this PR focus on the changes,
while the first commit just implements the output of the generator,
will make this easier to review.

Ulimately, I think a bespoke UI may make some of this management
easier, but we're trying not to focus on this sort of work at this
point.