Create Detections model and associated links
** Why are these changes being introduced:

* The application needs a way to record the outcome of the application's
  Detectors when they review the Terms we receive. These outcomes will
  be needed in a subsequent step, the application tries to place these
  Terms into the various Categories.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-33

** How does this address that need:

This defines a Detections model, which is a linkage between a Term and
a Detector. In order to allow the application's behavior to improve over
time, we also define a DETECTOR_VERSION value in the environment, which
will be updated as the application becomes more capable.

Detections are the unique combination of a Term, a Detector, for a given
DETECTOR_VERSION.

The Detection model includes several scopes, which are consulted during
the before_create lifecycle hook to prevent duplicate records from being
created. I've attempted to write tests to indicate how this works.

** Document any side effects to this change:

This change deviates from our earlier plans in two notable ways:

* The confidence value which we originally intended to store with the
  Detection has been abandoned. Because a Detector might be linked
  with multiple Categories, each with a differing confidence value,
  there is no way to record a Term-Detector confidence value without
  losing information. The confidence values will still be used during
  Categorization, but the values cannot be cached in the Detection
  records.

* The various record_* methods are no longer defined as part of the
  Detection class. Because a single Term will be joined to multiple
  Detection records, it made more sense for me to define the record_*
  methods on the Term class instead.