Switch to find_or_create_by after code review
Jeremy pointed out during code review that it might be worth using
find_or_create_by in the Term methods, rather than a combination of the
create method and a guard method in the Detection model.

This makes that change, adopting find_or_create_by inside Term and
removing the check_for_current method from the Detection model.

Along the way we update the specific assertion being looked for in the
Detection tests. Please note that there are tests in the Term model that
confirm the record_* methods an be called repeatedly without either
creating duplicate records, or throwing errors.