Move record methods into each Detector subclass
** Why are these changes being introduced:

The original proposal for implementing the Detection workflow involved
a series of record_* methods in the Term model, which then called
methods across the application to coordinate the creation of needed
records.

During code review, a more modular arrangement was suggested, where the
record_* methods would be distributed in the Detector subclasses, where
the work was actually being done. Those subclasses would then each
manage their own Detection record creation.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-33

** How does this address that need:

This makes that change, moving the work of recording Detections to the
Detector subclasses. The tests are refactored as well, to match this
arrangement.

** Document any side effects to this change:

Some of the record methods feel very close to being repetetive, which
makes me wonder whether we should add them to the base Detector class,
and allow overrides when a specific detector needs a different logic
(for example, the Journal and SuggestedResource look up against a list,
while StandardIdentifiers might provide its own regex-based record
method).