Updates the classes diagram in documentation
** Why are these changes being introduced:

There are some changes needed to the classes diagram, to keep things
current as we've built out this categorization process.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-83

** How does this address that need:

This updates the class diagram with the new methods, and new method
names, which have been added to the application. There were also some
misnamed classes ("DetectionCategory" instead of "DetectorCategory"),
which this fixes.

Additionally, this tries out namespaces in Mermaid to account for
class names like Detector::Journal, which haven't been included in the
diagram yet.

** Document any side effects to this change:

I may be overthinking this - but I'm finding this diagram to be very
helpful as we build out the application, and I'd like to try and keep
it as current as we can.