Build Categorization workflow
** Why are these changes being introduced:

With the Detection model now returning associated categories and
confidences, and the Categorization model ready to receive the outcome
of categorization, something needs to actually do the work.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-83

** How does this address that need:

This declares that the Term model will be responsible for performing the
Categorization work, pulling information from the Detections model,
shaping the response into something it can use, and overseeing the
creation of any needed Categorization records.

While each Detection record is returning its associated categories and
confidences, some Terms will end up tripping multiple Detectors - and
some of the associated Categories will end up being duplicated. Getting
all this information into a compact and processable structure is the
work of retrieve_detection_scores - which is private because nothing
should need to call that method except the calculate_categorization
method that manages all of this.

Contrarily, the work of combining all the individual detection
confidence values feels like something that needs to be testable - so
I've chosen to make that a public method for testing. Nothing else
should be calling the method, but I want to be transparent about what
caclulations we're doing.

I've tried to leave code comments in these methods to describe what the
data looks like at each step. This has been a pain to work through, and
I'm not sure I've done it right - so hopefully the comments will help
future-us.

I've tried to write the tests that seem warranted for this - but I'm
not sure that I've succeeded.

** Document any side effects to this change:

* I'm not 100% convinced that the public/private method choices I've
  made are the correct ones, and there may be aspects of this work that
  need tests as guardrails which I haven't written yet.