Add categorization information to GraphQL response
** Why are these changes being introduced:

Now that the logSearchEvent query is going to be initiating the
Detection and Categorization workflows, we can actually return the
results in the graphQL response for both lookupTerm and logSearchEvent.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-83

** How does this address that need:

This defines a Categories object type to the GraphQL vocabulary, and
includes an array of these objects to both the Term and SearchEvent
objects. The content of this Categories object spans both the category
name (from the Categories model) and the overall confidence value that
this term belongs to that category (the confidence value is part of the
Categorization model).

** Document any side effects to this change:

* This may be the first test that we've written for the lookupTerm
  function?