Add Categorization read-only dashboard
Why are these changes being introduced:

* It will be convenient to have a mechansim in-app for authenticated
  users to look at our Categorizations prior to us building out custom
  reports/dashboards

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TCO-93

How does this address that need:

* Adds a read-only administrate dashboard for Categorizations
* Updates ability.rb to allow all authenticated users to view the
  dashboard

Document any side effects to this change:

This introduces minor changes to how Term and SearchEvents displays
througout the Dashboards. Overall I think this is an improvement, but
there may be edge cases where this change is not desired.