Integrate Citation detector with Metrics
** Why are these changes being introduced:

Now that the Detection::Citation model is in place, we need to integrate
it with the existing Metrics reporting.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-107

** How does this address that need:

* Adds a citation field (integer) to the metrics_algorithms table

* Defines a match_citations method in the Algorithms model, to pass
  search events through the Citation detector and count those where one
  is found (via the detection? method). This also includes expanding the
  event_matches logic to account for this in the unmatched calculations.

* Creates a new search_event fixture, using the citation term fixture,
  to give the algorithm something to count.

* As a side effect, we have to tweak the isbn term fixture, because it
  was unintentionally tripping the citation as well. See the side effect
  section for a longer discussion of this.

* Adds tests to the Algorithm model to account for everything above,
  copying from other calculations.

** Document any side effects to this change:

The fact that this needed to tweak a different term fixture is an
indication that our detectors are starting to overlap. This is a real
phenomenon in production, and is expected as we start to build more
detectors.

I do not have a notion for now about how to address this in the more
constrained test environment, where we can carefully craft fixtures to
do exactly what we intend. It feels like this behavior should be
confirmed via explicit tests that address this overlap, but the tests
we have so far are focused on one-and-only-one detector, so I feel
okay about tweaking things to keep them separate.

We do have a term fixture that trips multiple detectors, confirming
that they get counted in all applicable categories, so we may be
covered already?