Create Validations model and associations
** Why are these changes being introduced:

We need a Validations model that will be attached to the actions taken
by TACOS (the Detections and Categorizations that are created during
normal operation).

Validation records will allow library staff to review and give feedback
about the actions taken by the application. This feedback will be used
to improve the operation of the tool, as well as to guide future
development.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-8

** How does this address that need:

This defines the migration to create the Validations table, and updates
the references to it from related classes. It also runs `make annotate`
to keep some minimal documentation going.

The model itself has no methods, and I'm not sure yet how it will be
flushed out. The most notable aspect is that it has a polymorphic
association with _both_ the Detections and Categorizations models, via
the :validatable construct, because all of those records will need to be
given feedback. The (probably poorly named) "judgement" field is meant
to store the postive / negative feedback about the action that TACOS
took.

** Document any side effects to this change:

Because TACOS only records positive actions (detections that were made,
and categorization scores that were calculated), this model currently
can only provide negative feedback (this detection should not have
fired, or this categorization should not have been performed). There
is not yet a way to say "this detector should have fired" or "this
category should have been assigned."

Further, there is a slight awkwardness approaching because the UI is
going to focus on the Term, but that term is potentially subject to
many Detections and Categorizations - so the form in the next commit
is going to need to take one input and spawn multiple Validation
records.