Creates basic read-only validation UI
** Why are these changes being introduced:

As part of allowing users to validate the actions taken by TACOS, we
need to first create an interface showing what those actions have been.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-8

** How does this address that need:

This commit creates a very basic interface to present what actions have
been taken with regard to a single term.

There is an index page, which shows all terms that have prompted any
activity (not all terms in the system).

From this index, the user can click into a display for any given term,
which will list all the detectors and categories in the system, with an
indication of which detectors activated for that term. The final score
for each category is also displayed.

There is not yet a form on this page, and the details of the UI need to
be more closely considered. This is a starting point, not the end.

** Document any side effects to this change:

This may be a side effect, or it may not, but: I've added the manage
role for Validations to the basic user, rather than creating a specific
role for this work. This is in line with what we've done so far, but I
am nervous about relying on basic roles for too much.