Update documentation
** Why are these changes being introduced:

We are adding a Confirmation class for users to provide ground truth for
how different Terms should be categorized. The first step is to update
our documentation to work out how this is going to work.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-8

** How does this address that need:

This updates the class diagram and workflow documentation to include the
Confirmation class.

* User and Confirmation are now grouped in the class diagram, with a
common visual language.

* The categorization workflow document is updated to use the new
vocabulary "confirmation" rather than our earlier "validation".

** Document any side effects to this change:

As this commit only changes documentation, there are no side effects
yet.