Create UserCategorization model
** Why are these changes being introduced:

With the documentation updated, we need to actually change the models.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-8

** How does this address that need:

* Adds the Confirmation model, with relationships to User, Term, and
Category (optionally). The category relationship is optional because
there is also the option for a user to flag the term for removal. This
flag is stored on the Confirmation model so that we have a record of who
flagged it, in case we want to follow up with them. The implication is
that we will review these flags, not perform the removal immediately.

There is a uniqueness constraint on the Confirmation model, to limit
users from confirming the same term twice.

* Two scopes are added to the Term model which filter records that have
(and have not) been confirmed by a user.

* A fourth category is added to the database seeds - "Undefined". The
automated categorization does not interact with this category, but users
have the option to place a term there in their workflow. The UI which is
coming in the next commit then will provide five options for manual
confirmation: one for each category, and then a "flag" option.

* Tests and fixtures are added to support and confirm all of the above.

** Document any side effects to this change:

There are a few side effects here:

* While adding the 'flagged' field makes sense to me on the Confirmation
model, it may also need to be added to the Term model (or at least add
a scope to grab flagged records)

* The existing scopes on the Term model do not pay attention to which
user has confirmed the term - so at the moment getting multiple
confirmations from multiple users may not be possible without further
work.

* Because we are changing the database seeds, we will need to take care
when promoting this change ito production.