Add automatic fingerprinting for terms
** Why are these changes being introduced:

After watching our search traffic for a few months, we have seen some
terms come in that are clearly related:

* 'Scientific American'
* 'scientific american'
* '&quot;Scientific american&quot;'

While there are valid reasons to store the term exactly as the user has
submitted it - there are also good reasons to standardize these values,
to look for clusters and related terms.

** Relevant ticket(s):

n/a

** How does this address that need:

This defines a TermFingerprint model, which is related to the Term model
via a belongs_to relationship (a Term belongs to its TermFingerprint,
and a TermFingerprint can have many Terms). The migration to define this
model includes a migration to populate records for Terms we've already
received.

The Term model gets two lifecycle hooks, one to add new fingerprints for
every term, and another to delete a TermFingerprint if no remaining
Terms reference it.

Beyond the methods needed to handle these creations and deletions in as
seemless a way as possible, we also add a .cluster method to the Term
model, which will return an array of all related Terms (but not the Term
itself - so a Term that has a unique fingerprint will return an empty
array). This method will be used as part of a future inspection UI.

Building the models in this way will also allow for querying based on
shared fingerprints - for example by adding a :has_many_terms scope
on the TermFingerprint model.

** Document any side effects to this change:

Two things:

1. A quirk of this implementation is that it is possible to delete a
TermFingerprint record, which does not delete the related Terms (and
thus also SearchEvents, Detections, and Categorizations). The Term will
then have a null fingerprint. The next time the Term is saved, its
fingerprint will be regenerated. No other operation should be impacted
by this arrangement.

For the life of me I can't anticipate why we might delete a fingerprint
- but the relationship needs to be optional to avoid further pain when
working in the console.

2. The process for calculating this TermFingerprint value is similar -
but not identical - to that for handling the SuggestedResource records.
The difference is that the TermFingerprint method removes "&quot;"
sequences - which might need to be added to SuggestedResource. Both
methods should probably be abstracted out to a shared helper method,
honestly.