Finish implementing confirmation workflow
** Why are these changes being introduced:

A previous PR started the implementation of our confirmation workflow,
but did not include a form in the view template, nor the controller
logic to receive the form submissions.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/tco-101

** How does this address that need:

This finishes the job of the previous PR, completing the confirmation
workflow. More specifically, it:

* Updates the routes information, moving the "index" display to a better
route name, and defining Confirmations as resources under Terms. This
makes dealing with routing more logical, and takes advantage of Rails'
magic. Only the new and create methods are implemented.

* Moves the confirmation form out of the Terms controller, and into a
new Confirmation controller, which is where records like this should be
managed, in the Confirmation#new method.

* Form submissions are received in Confirmation#create, which is helped
by two new private methods for dealing with category values (which are
either a category_id or a "flag" boolean), and for user feedback based
on the create result.

* The site_nav and tests are updated to reflect the new controller and
route information

* The view templates are updated, most notably the new confirmation
template, which now has a working form using Rails' built-in form tool.

** Document any side effects to this change:

The confirmation form has some inline styles to quickly get the UI into
something not-awful.

Rubocop is complaining about lack of translation in the feedback
messages in the Confirmation controller,

The final else clause in the Confirmation#feedback_for method is not
something I can figure out a test for - but I'm loathe to not have
a default clause in that conditional. The .save method can return
false, but usually that means an error message, which I've provided
a catch for the most likely scenario (and a test showing how it gets
produced).

I'm not sure about one of the confirmation controller tests - the one
confirming that terms disappear from the list after a confirmation is
created. We already have a test for the scope that provides this list,
but it felt appropriate to try and test at the controller level as
well? I don't remember writing tests like this in the past, though -
so maybe it isn't relevant here.